chore: rename domain_fn to origin_fn
