+++ packages/driver/src/cross-origin/cypress.ts
@@ -10,7 +10,7 @@ import { $Location } from '../cypress/location'
 import $Commands from '../cypress/commands'
 import { create as createLog } from '../cypress/log'
 import { bindToListeners } from '../cy/listeners'
-import { handleOriginFn } from './domain_fn'
+import { handleOriginFn } from './origin_fn'
 import { FINAL_SNAPSHOT_NAME } from '../cy/snapshots'
 import { handleLogs } from './events/logs'
 import { handleSocketEvents } from './events/socket'
